make factors floats

+++ pool/config.go
@@ -45,11 +45,11 @@ type Config struct {
 // EffectiveGasPrice has parameters for the effective gas price calculation.
 type EffectiveGasPrice struct {
 	// L1GasPriceFactor is the percentage of the L1 gas price that will be used as the L2 min gas price
-	L1GasPriceFactor uint64 `mapstructure:"L1GasPriceFactor"`
+	L1GasPriceFactor float64 `mapstructure:"L1GasPriceFactor"`
 
 	// ByteGasCost is the gas cost per byte
 	ByteGasCost uint64 `mapstructure:"ByteGasCost"`
 
 	// MarginFactor is the margin factor percentage to be added to the L2 min gas price
-	MarginFactor uint64 `mapstructure:"MarginFactor"`
+	MarginFactor float64 `mapstructure:"MarginFactor"`
 }
